requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.10.5
click==8.3.0
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
dlib==20.0.0
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.1
idna==3.11
ipykernel==7.1.0
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.9.1
kiwisolver==1.4.9
lightning-utilities==0.15.2
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mpmath==1.3.0
multidict==6.7.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn==8.2.0.51
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
opencv-python-headless==4.12.0.88
packaging==25.0
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.5
PyYAML==6.0.3
pyzmq==27.1.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
stringzilla==4.2.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==1.0.21
torch==2.9.0
torchmetrics==1.8.2
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
wcwidth==0.2.14
Werkzeug==3.1.3
yarl==1.22.0
